requirements.txt
# ==========================================
# RIS Scraper System - Dependencies
# ==========================================

# Core
python-dotenv==1.0.0

# Web Scraping
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# OCR for CAPTCHA
Pillow==10.1.0
ddddocr>=1.4.0

# Data Processing
pandas==2.1.4

# Database
psycopg2-binary==2.9.9

# API Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
jinja2==3.1.2
aiofiles==23.2.1

# HTTP Client
requests==2.31.0

# Scheduler
APScheduler>=3.10.0

# Logging (Loki)
python-logging-loki>=0.3.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0
